Moves application components under an app/ directory structure, so we can organize our code a bit better and make room for future separation like a state/ directory.
