+++ client/extensions/woocommerce/app/dashboard/index.js
