+++ client/extensions/woocommerce/app/products/product-variation-types-form.js
